backend/pyproject.toml
[project]
name = "edgecase-backend"
version = "0.1.0"
description = "Edgecase - Supply Chain Analyser for Risk and Overconfidence"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "neo4j>=5.16.0",
    "qdrant-client>=1.7.0",
    "redis>=5.0.1",
    
    # HTTP & Scraping
    "httpx>=0.26.0",
    "aiohttp>=3.9.1",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "lxml_html_clean",
    "playwright>=1.41.0",
    "newspaper3k>=0.2.8",
    "duckduckgo-search>=4.1.1",
    "undetected-chromedriver",
    
    # NLP & AI
    "spacy>=3.7.2",
    "openai>=1.12.0",
    "numpy>=1.26.3",
    "pandas>=2.1.4",
    "scikit-learn>=1.4.0",
    
    # Utilities
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "celery>=5.3.6",
    "nest-asyncio>=1.6.0",
    
    # Reddit API
    "praw>=7.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "ruff>=0.1.11",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
